Make changes required from the pr

+++ db/migrate/20220517173525_create_foods.rb
@@ -3,7 +3,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.string :measurement_unit
-      t.money :price
+      t.float :price
       t.references :user, null: false, foreign_key: true, index: true
 
       t.timestamps
